day_14/requirements.txt
faiss-cpu
# Day 14 - Production RAG Pipeline Requirements

# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.4.0

# RAG Pipeline dependencies (from Day 12)
sentence-transformers>=2.2.0
chromadb>=0.4.0
openai>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Production dependencies
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and metrics
prometheus-client>=0.18.0
structlog>=23.1.0

# Security and middleware
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP client
httpx>=0.25.0
aiofiles>=23.2.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
mypy>=1.6.0

# Optional: Alternative model providers
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.1.0

# System utilities
psutil>=5.9.0
